+++ src/main/java/com/atlauncher/gui/dialogs/EditModsDialog.java
@@ -19,6 +19,7 @@ package com.atlauncher.gui.dialogs;
 
 import java.awt.BorderLayout;
 import java.awt.Dimension;
+import java.awt.FlowLayout;
 import java.awt.event.ItemEvent;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
@@ -30,6 +31,7 @@ import java.util.stream.Collectors;
 
 import javax.swing.AbstractButton;
 import javax.swing.JButton;
+import javax.swing.JCheckBox;
 import javax.swing.JDialog;
 import javax.swing.JLabel;
 import javax.swing.JList;
@@ -66,6 +68,7 @@ public class EditModsDialog extends JDialog {
     private JScrollPane scroller1, scroller2;
     private JButton addButton, addCurseModButton, checkForUpdatesButton, enableButton, disableButton, removeButton,
             closeButton;
+    private JCheckBox selectAllEnabledModsCheckbox, selectAllDisabledModsCheckbox;
     private JLabel topLabelLeft, topLabelRight;
     private ArrayList<ModsJCheckBox> enabledMods, disabledMods;
 
@@ -139,13 +142,41 @@ public class EditModsDialog extends JDialog {
         labels.setEnabled(false);
         split.setRightComponent(labels);
 
+        JPanel topLeftPanel = new JPanel(new FlowLayout());
+
         topLabelLeft = new JLabel(GetText.tr("Enabled Mods"));
         topLabelLeft.setHorizontalAlignment(SwingConstants.CENTER);
-        labels.setLeftComponent(topLabelLeft);
+        topLeftPanel.add(topLabelLeft);
+
+        selectAllEnabledModsCheckbox = new JCheckBox();
+        selectAllEnabledModsCheckbox.addActionListener(e -> {
+            boolean selected = selectAllEnabledModsCheckbox.isSelected();
+
+            enabledMods.stream().forEach(em -> {
+                em.setSelected(selected);
+            });
+        });
+        topLeftPanel.add(selectAllEnabledModsCheckbox);
+
+        labels.setLeftComponent(topLeftPanel);
+
+        JPanel topRightPanel = new JPanel(new FlowLayout());
 
         topLabelRight = new JLabel(GetText.tr("Disabled Mods"));
         topLabelRight.setHorizontalAlignment(SwingConstants.CENTER);
-        labels.setRightComponent(topLabelRight);
+        topRightPanel.add(topLabelRight);
+
+        selectAllDisabledModsCheckbox = new JCheckBox();
+        selectAllDisabledModsCheckbox.addActionListener(e -> {
+            boolean selected = selectAllDisabledModsCheckbox.isSelected();
+
+            disabledMods.stream().forEach(dm -> {
+                dm.setSelected(selected);
+            });
+        });
+        topRightPanel.add(selectAllDisabledModsCheckbox);
+
+        labels.setRightComponent(topRightPanel);
 
         modsInPack = new JSplitPane();
         modsInPack.setDividerLocation(275);
@@ -372,6 +403,11 @@ public class EditModsDialog extends JDialog {
                 || (enabledMods.size() != 0 && enabledMods.stream().anyMatch(AbstractButton::isSelected)));
         enableButton.setEnabled(disabledMods.size() != 0 && disabledMods.stream().anyMatch(AbstractButton::isSelected));
         disableButton.setEnabled(enabledMods.size() != 0 && enabledMods.stream().anyMatch(AbstractButton::isSelected));
+
+        selectAllEnabledModsCheckbox
+                .setSelected(enabledMods.size() != 0 && enabledMods.stream().allMatch(AbstractButton::isSelected));
+        selectAllDisabledModsCheckbox
+                .setSelected(disabledMods.size() != 0 && disabledMods.stream().allMatch(AbstractButton::isSelected));
     }
 
     private void checkForUpdates() {