+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Fix mods added via the button not getting mod information
 - Fix adding multiple mods via drag and drop asking for type every file
 - Fix closing add mod file dialog still addding the file
+- Add select all to the edit mods dialog